data_analysis/requirements.txt
flask
pandas
numpy
gunicorn
matplotlib
seaborn
requests
python-dotenv